Add files via upload
Companies need a workforce of diverse talents and backgrounds to succeed in an increasingly complex and heterogeneous world. Diversity and inclusion are business imperatives, not just nice-to-haves. But actually achieving this and unlocking its potential involves a whole set of practical challenges.
Calculating the following measures could help to define proper KPIs:

# of men
# of women
# of leavers
% employees promoted (FY21)
% of women promoted
% of hires men
% of hires women
% turnover 
Average performance rating: men
Average Performance rating: women